PAM-3682: Forenklar ved å gjenbruke side effect.
