+++ src/common/hooks/index.js
@@ -1,6 +1,6 @@
 /* eslint-disable import/prefer-default-export */
 import useDocumentTitle from './useDocumentTitle';
-import useGoogleAnalytics from './useGoogleAnalytics';
+import useTrackPageview from './useTrackPageview';
 import useScrollToTop from './useScrollToTop';
 
-export { useDocumentTitle, useGoogleAnalytics, useScrollToTop };
+export { useDocumentTitle, useTrackPageview, useScrollToTop };
