--- src/common/hooks/useGoogleAnalytics.js
@@ -1,14 +1,0 @@
-/* eslint-disable no-undef */
-import { useEffect } from 'react';
-
-export default (page, title) => {
-    useEffect(() => {
-        try {
-            ga('set', 'page', page);
-            ga('set', 'title', title);
-            ga('send', 'pageview');
-        } catch (e) {
-            // Google Analytics er ikke definert (dette skjer f.eks. om en bruker blokkerer tracking).
-        }
-    }, []);
-};
