+++ src/search/Search.js
@@ -29,7 +29,7 @@ import SearchResultCount from './searchResults/SearchResultCount';
 import SearchResults from './searchResults/SearchResults';
 import ShowResultsButton from './showResultsButton/ShowResultsButton';
 import Sorting from './sorting/Sorting';
-import { useDocumentTitle, useGoogleAnalytics } from '../common/hooks';
+import { useDocumentTitle, useTrackPageview } from '../common/hooks';
 
 const Search = ({
     initialSearch,
@@ -45,7 +45,7 @@ const Search = ({
         restoreStateFromUrl();
     });
 
-    useGoogleAnalytics(CONTEXT_PATH, 'Ledige stillinger');
+    useTrackPageview(CONTEXT_PATH, 'Ledige stillinger');
 
     useEffect(() => {
         const doNotOverrideDocumentTitle = isPredefinedDocumentTitle(document.title);
